+++ js/data.js
@@ -1,0 +1,33 @@
+export const NAMES = [
+  'Иван',
+  'Мария',
+  'Алексей',
+  'Ольга',
+  'Дмитрий',
+  'Анна',
+  'Сергей',
+  'Елена',
+  'Артём',
+  'Наталья'
+];
+export const DESCRIPTIONS = [
+  'Прекрасный закат на море',
+  'Горный пейзаж',
+  'Уютный вечер в городе',
+  'Прогулка по лесу',
+  'Архитектурный шедевр',
+  'Момент радости',
+  'Путешествие мечты',
+  'Уличное искусство',
+  'Кофе утром',
+  'Зимняя сказка'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
